+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Add exclude directories option to aggregateTranslations script and cli
 
 4.20.0 - (December 6, 2018)
 ----------